Completely remade command handler
