+++ Modules/Bot.ts
@@ -1,5 +1,5 @@
-import { CommandClient, Message, Member } from 'eris';
-import { Points, Upvote } from './index';
+import { CommandClient } from 'eris';
+import { commands } from './Commands';
 import { config } from '../config';
 
 const commandOptions = {
@@ -20,69 +20,6 @@ bot.on('error', (e: Error) => {
     console.log(e);
 });
 
-// Commands
-
-bot.registerCommand('selfReminder', (msg: Message): string => {
-    const reminder = Upvote.setReminder(msg.member!);
-
-    if (reminder) {
-        return 'Reminder set. I will remind you to upvote in 12 hours from now.';
-    } else {
-        return 'You already have a reminder set.';
-    }
-}, {
-    description: 'Sets an upvote reminder for yourself.',
-    fullDescription: 'Sets a new upvote reminder for yourself. Doesn\'t work if there is already a reminder set for you.',
-});
-
-// Ruler Only Commands
-
-bot.registerCommand('pointsChange', async (msg: Message, args: string[]) => {
-    const [ userID, amount ] = args;
-
-    if (!userID || !amount) {
-        return 'Not enough arguments.';
-    } else if (!+amount) {
-        return 'Amount is not a number.';
-    }
-
-    const userPoints = await Points.handle(userID, +amount);
-
-    return `<@${userID}>'s points are now at ${userPoints}`;
-}, {
-    requirements: {
-        roleIDs: ['378293035852890124'],
-    },
-    description: '`**(Ruler only)**` Changes points by given amount.',
-    fullDescription: '`**(Ruler only)**`\nIncrements or decrements the points of the given user by given amount.',
-    usage: '`userID` `amount`',
-});
-
-bot.registerCommand('newReminder', (msg: Message, args: string[]): string => {
-    const ease = bot.guilds.get('365236789855649814');
-    const user: Member | undefined = ease!.members.find((u: any) => u.id === args[0]);
-
-    if (!user) {
-        return 'Couldn\'t find that user.';
-    } else if (!args || args.length > 2) {
-        return 'Not enough or too many arguments.';
-    } else if (bot.user.id === args[0]) {
-        return 'I don\'t need to remind myself to upvote, thanks.';
-    }
-
-    const reminder = Upvote.setReminder(user);
-
-    if (reminder) {
-        return 'Reminder set.';
-    } else {
-        return 'This user already has a reminder set.';
-    }
-
-}, {
-    requirements: {
-        roleIDs: ['378293035852890124'],
-    },
-    description: '`**(Ruler only)**` Sets an upvote reminder for a user.',
-    fullDescription: '`**(Ruler only)**`\nSets a new upvote reminder for the given user. Doesn\'t work if there is already a reminder set.',
-    usage: '`userID`',
+commands.forEach((element) => {
+    bot.registerCommand(element.label, element.execute, element.options)
 });
