+++ tslint.json
@@ -11,7 +11,8 @@
         "max-line-length": false,
         "object-literal-shorthand": false,
         "no-console": false,
-        "object-literal-sort-keys": false
+        "object-literal-sort-keys": false,
+        "interface-name": false,
     },
     "rulesDirectory": []
 }
